US-6930: patch script check if reconcile still exists

+++ bin/addons/account_reconciliation/account_move_line.py
@@ -756,7 +756,7 @@ class account_move_unreconcile(osv.osv):
                 netsvc.LocalService("workflow").trg_validate(uid, 'account.invoice', invoice_reopen, 'open_test', cr)
         elif context.get('sync_update_execution') and vals.get('reconcile_sdref'):
             rec_id = self.pool.get('account.move.reconcile').find_sd_ref(cr, uid, vals.get('reconcile_sdref'), context=context)
-            if rec_id:
+            if rec_id and self.pool.get('account.move.reconcile').exists(cr, uid, rec_id, context=context):
                 self.pool.get('account.move.reconcile').unlink(cr, uid, rec_id, context=context)
         return super(account_move_unreconcile, self).create(cr, uid, vals, context)
